+++ ignite/services/scaffolder/init.go
@@ -26,7 +26,7 @@ func Init(
 	tracer *placeholder.Tracer,
 	root, name, addressPrefix string,
 	noDefaultModule, skipGit, minimal bool,
-	params []string,
+	params, moduleConfigs []string,
 ) (path string, err error) {
 	pathInfo, err := gomodulepath.Parse(name)
 	if err != nil {
@@ -46,7 +46,7 @@ func Init(
 	path = filepath.Join(root, appFolder)
 
 	// create the project
-	err = generate(ctx, tracer, pathInfo, addressPrefix, path, noDefaultModule, minimal, params)
+	err = generate(ctx, tracer, pathInfo, addressPrefix, path, noDefaultModule, minimal, params, moduleConfigs)
 	if err != nil {
 		return "", err
 	}
@@ -74,10 +74,16 @@ func generate(
 	addressPrefix,
 	absRoot string,
 	noDefaultModule, minimal bool,
-	params []string,
+	params, moduleConfigs []string,
 ) error {
 	// Parse params with the associated type
 	paramsFields, err := field.ParseFields(params, checkForbiddenTypeIndex)
+	if err != nil {
+		return err
+	}
+
+	// Parse configs with the associated type
+	configsFields, err := field.ParseFields(moduleConfigs, checkForbiddenTypeIndex)
 	if err != nil {
 		return err
 	}
@@ -129,6 +135,7 @@ func generate(
 			AppName:    pathInfo.Package,
 			AppPath:    absRoot,
 			Params:     paramsFields,
+			Configs:    configsFields,
 			IsIBC:      false,
 		}
 		// Check if the module name is valid